+++ src/Components/Common/LoginHeader/UserProfile.tsx
@@ -1,39 +1,8 @@
 import React from "react";
-import { useQuery } from "@tanstack/react-query";
-import fetchUserProfile from "@/apis/user/user";
+
 import LoginHeaderDropdown from "./LoginHeaderDropdown";
 
-interface UserProfileType {
-  id: number;
-  email: string;
-  nickname: string;
-  profileImageUrl: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-const UserProfile: React.FC = ({}) => {
-  const {
-    data: UserProfile,
-    isLoading,
-    isError,
-  } = useQuery<UserProfileType>({
-    queryKey: ["userProfile"],
-    queryFn: () => fetchUserProfile(),
-  });
-
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-
-  if (isError) {
-    return <div>Error loading profile</div>;
-  }
-
-  if (!UserProfile) {
-    return <div></div>;
-  }
-
+const UserProfile = () => {
   return (
     <div className="flex items-center justify-center text-gnDarkBlack">
       <button>
@@ -41,10 +10,6 @@ const UserProfile: React.FC = ({}) => {
       </button>
       <div className="mx-4 h-6 w-px bg-gnGray300"></div>
       <LoginHeaderDropdown />
-      <button className="flex items-center justify-between gap-3">
-        <img src={UserProfile.profileImageUrl} alt="Profile Picture" />
-        {UserProfile.nickname}
-      </button>
     </div>
   );
 };
